Add protobuf support for parsing/encoding link costs in StripedLocatedBlock
This commit adds necessary changes to .proto files and helper functions
to allow the encoding & parsing of link costs added to the
StripedLocatedBlock object on the namenode side.

Without the encoding support, the namenode won't send the link costs in
its RPC response. Without the parsing support, the client will be unable
to parse the received link cost values in the RPC response