Add simple test for DFSStripedInputStream GDA changes
The testcase added is very similar to testPreadWithDnFailure in
TestDFSStripedInputStream.java. That test verifies correct decoding when
a DN is shut down.

This testcase verifies similar behaviour. The only difference is that
one DN is remote and has a very high link cost. The greedy EC read
mechanism should essentially treat this DN as a failed DN, and thus this
test should perform similarly to the other test.